requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
matplotlib>=3.5.0
opencv-python>=4.5.0
tensorboard>=2.8.0
tqdm>=4.62.0
Pillow>=9.0.0
ultralytics>=8.0.0  
pytorch-fid>=0.3.0  
scikit-image>=0.19.0
huggingface_hub>=0.17.0 